pyproject.toml
[tool.poetry]
name = "Shop_API"
version = "0.1.0"
description = "Test Assignment"
authors = ["alekspletnev <alekspletnev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
